fix: 将token_scores转换为float类型后再生成热力图

+++ test/screenspot_v2/visualization_token.py
@@ -158,7 +158,7 @@ def visualize_token_scores(
     image_copy = image.copy()
 
     # 将token分数转换为热力图
-    scores = token_scores.cpu().numpy()
+    scores = token_scores.float().cpu().numpy()
     heatmap = scores.reshape(num_patches_h, num_patches_w)
 
     # 调整热力图大小以匹配图像
